Refactored json utilities into a json module Added missing implementation for utf-16
