+++ superset/sqllab/execution_context_convertor.py
@@ -19,11 +19,9 @@ from __future__ import annotations
 import logging
 from typing import Any, TYPE_CHECKING
 
-import simplejson as json
-
-import superset.utils.core as utils
 from superset.sqllab.command_status import SqlJsonExecutionStatus
 from superset.sqllab.utils import apply_display_max_row_configuration_if_require
+from superset.utils import json as json_utils
 
 logger = logging.getLogger(__name__)
 
@@ -52,23 +50,16 @@ class ExecutionContextConvertor:
 
     def serialize_payload(self) -> str:
         if self._exc_status == SqlJsonExecutionStatus.HAS_RESULTS:
-            sql_results = apply_display_max_row_configuration_if_require(
-                self.payload, self._max_row_in_display_configuration
+            return json_utils.dumps(
+                apply_display_max_row_configuration_if_require(
+                    self.payload, self._max_row_in_display_configuration
+                ),
+                default=json_utils.pessimistic_json_iso_dttm_ser,
+                ignore_nan=True,
             )
-            try:
-                return json.dumps(
-                    sql_results,
-                    default=utils.pessimistic_json_iso_dttm_ser,
-                    ignore_nan=True,
-                )
-            except UnicodeDecodeError:
-                return json.dumps(
-                    sql_results,
-                    default=utils.pessimistic_json_iso_dttm_ser,
-                    ensure_ascii=False,
-                    ignore_nan=True,
-                )
 
-        return json.dumps(
-            {"query": self.payload}, default=utils.json_int_dttm_ser, ignore_nan=True
+        return json_utils.dumps(
+            {"query": self.payload},
+            default=json_utils.json_int_dttm_ser,
+            ignore_nan=True,
         )
