+++ superset/views/api.py
@@ -33,7 +33,7 @@ from superset.commands.chart.exceptions import (
 from superset.legacy import update_time_range
 from superset.models.slice import Slice
 from superset.superset_typing import FlaskResponse
-from superset.utils import core as utils
+from superset.utils import json as json_utils
 from superset.utils.date_parser import get_since_until
 from superset.views.base import api, BaseSupersetView, handle_api_exception
 
@@ -73,8 +73,8 @@ class Api(BaseSupersetView):
         query_context.raise_for_access()
         result = query_context.get_payload()
         payload_json = result["queries"]
-        return json.dumps(
-            payload_json, default=utils.json_int_dttm_ser, ignore_nan=True
+        return json_utils.dumps(
+            payload_json, default=json_utils.json_int_dttm_ser, ignore_nan=True
         )
 
     @event_logger.log_this
