+++ superset/views/chart/views.py
@@ -22,7 +22,7 @@ from superset import security_manager
 from superset.constants import MODEL_VIEW_RW_METHOD_PERMISSION_MAP, RouteMethod
 from superset.models.slice import Slice
 from superset.superset_typing import FlaskResponse
-from superset.utils import core as utils
+from superset.utils import json as json_utils
 from superset.views.base import DeleteMixin, DeprecateModelViewMixin, SupersetModelView
 from superset.views.chart.mixin import SliceMixin
 
@@ -41,10 +41,10 @@ class SliceModelView(
     method_permission_name = MODEL_VIEW_RW_METHOD_PERMISSION_MAP
 
     def pre_add(self, item: "SliceModelView") -> None:
-        utils.validate_json(item.params)
+        json_utils.validate_json(item.params)
 
     def pre_update(self, item: "SliceModelView") -> None:
-        utils.validate_json(item.params)
+        json_utils.validate_json(item.params)
         security_manager.raise_for_ownership(item)
 
     def pre_delete(self, item: "SliceModelView") -> None:
