+++ superset/views/core.py
@@ -71,10 +71,9 @@ from superset.models.slice import Slice
 from superset.models.sql_lab import Query
 from superset.models.user_attributes import UserAttribute
 from superset.superset_typing import FlaskResponse
-from superset.utils import core as utils
+from superset.utils import core as utils, json as json_utils
 from superset.utils.cache import etag_cache
 from superset.utils.core import (
-    base_json_conv,
     DatasourceType,
     get_user_id,
     ReservedUrlParameters,
@@ -576,8 +575,8 @@ class Superset(BaseSupersetView):
 
         return self.render_template(
             "superset/basic.html",
-            bootstrap_data=json.dumps(
-                bootstrap_data, default=utils.pessimistic_json_iso_dttm_ser
+            bootstrap_data=json_utils.dumps(
+                bootstrap_data, default=json_utils.pessimistic_json_iso_dttm_ser
             ),
             entry="explore",
             title=title,
@@ -753,7 +752,7 @@ class Superset(BaseSupersetView):
             )
 
         return json_success(
-            json.dumps(
+            json_utils.dumps(
                 [
                     {
                         "slice_id" if key == "chart_id" else key: value
@@ -765,7 +764,7 @@ class Superset(BaseSupersetView):
                     }
                     for slc in slices
                 ],
-                default=base_json_conv,
+                default=json_utils.base_json_conv,
             ),
         )
 
@@ -814,12 +813,12 @@ class Superset(BaseSupersetView):
             "superset/spa.html",
             entry="spa",
             title=dashboard.dashboard_title,  # dashboard title is always visible
-            bootstrap_data=json.dumps(
+            bootstrap_data=json_utils.dumps(
                 {
                     "user": bootstrap_user_data(g.user, include_perms=True),
                     "common": common_bootstrap_payload(),
                 },
-                default=utils.pessimistic_json_iso_dttm_ser,
+                default=json_utils.pessimistic_json_iso_dttm_ser,
             ),
             standalone_mode=ReservedUrlParameters.is_standalone_mode(),
         )
@@ -919,8 +918,8 @@ class Superset(BaseSupersetView):
         return self.render_template(
             "superset/spa.html",
             entry="spa",
-            bootstrap_data=json.dumps(
-                payload, default=utils.pessimistic_json_iso_dttm_ser
+            bootstrap_data=json_utils.dumps(
+                payload, default=json_utils.pessimistic_json_iso_dttm_ser
             ),
         )
