+++ superset/views/sql_lab/views.py
@@ -28,7 +28,7 @@ from superset import db
 from superset.constants import MODEL_VIEW_RW_METHOD_PERMISSION_MAP, RouteMethod
 from superset.models.sql_lab import Query, SavedQuery, TableSchema, TabState
 from superset.superset_typing import FlaskResponse
-from superset.utils import core as utils
+from superset.utils import json as json_utils
 from superset.utils.core import get_user_id
 from superset.views.base import (
     BaseSupersetView,
@@ -140,7 +140,7 @@ class TabStateView(BaseSupersetView):
         if tab_state is None:
             return Response(status=404)
         return json_success(
-            json.dumps(tab_state.to_dict(), default=utils.json_iso_dttm_ser)
+            json.dumps(tab_state.to_dict(), default=json_utils.json_iso_dttm_ser)
         )
 
     @has_access_api
