+++ superset/viz.py
@@ -66,7 +66,7 @@ from superset.superset_typing import (
     VizData,
     VizPayload,
 )
-from superset.utils import core as utils, csv
+from superset.utils import core as utils, csv, json as json_utils
 from superset.utils.cache import set_and_log_cache
 from superset.utils.core import (
     apply_max_row_limit,
@@ -440,8 +440,8 @@ class BaseViz:  # pylint: disable=too-many-public-methods
 
     @deprecated(deprecated_in="3.0")
     def get_json(self) -> str:
-        return json.dumps(
-            self.get_payload(), default=utils.json_int_dttm_ser, ignore_nan=True
+        return json_utils.dumps(
+            self.get_payload(), default=json_utils.json_int_dttm_ser, ignore_nan=True
         )
 
     @deprecated(deprecated_in="3.0")
@@ -641,9 +641,9 @@ class BaseViz:  # pylint: disable=too-many-public-methods
     @staticmethod
     @deprecated(deprecated_in="3.0")
     def json_dumps(query_obj: Any, sort_keys: bool = False) -> str:
-        return json.dumps(
+        return json_utils.dumps(
             query_obj,
-            default=utils.json_int_dttm_ser,
+            default=json_utils.json_int_dttm_ser,
             ignore_nan=True,
             sort_keys=sort_keys,
         )
